% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/BayesConsensus.R
\name{BayesConsensus}
\alias{BayesConsensus}
\title{A Cat Function}
\usage{
BayesConsensus(AnswersGiven, Competancy, numAnswers, Prior = -1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

