% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CleanFreelist.R
\name{CleanFreeList}
\alias{CleanFreeList}
\title{CleanFreeList}
\usage{
CleanData<-CleanFreeList(mydata, Order="Order",Subj="Subj",CODE="CODE",ejectBadSubj=T,deleteDoubleCode=T,ConsolidateOrder=T,RemoveMissingData=T)
}
\arguments{
\item{mydata}{The Freelist data. This should be a data frame, where each row contains a single response from a single respondent. For each such response, you need to know the subj number (or some form of unique identifier), the response (or "CODE" of the response **Explain more??**) and the the ranking/order of the response (What was this respondents the first response? The second? The Seventeenth?)}

\item{Order}{This is the name of the column which contains the "Order" information. For each subject responses should be ordered uniquely from 1 to N, where N is the number of responses. There should be no gaps or double ups. Defaults to "Order"}

\item{Subj}{This is the name of the column containing your subject names/numbers. Each subject should have a unique identifier in this column (Is this confusing, given that each subject may take several rows?). Defaults to "Subj".}

\item{CODE}{This is the name of the column containing your subject names/numbers. Each subject should have a unique identifier in this column (Is this confusing, given that each subject may take several rows?). Defaults to "Subj".}

\item{ejectBadSubj}{Do you want to eject all subjects who have (for whatever reason) bad data? For example duplicates or missing order entries. defaults to true.}

\item{deleteDoubleCode}{If someone says "Apple" twice, do you want to drop all but the first instance? defaults to true.}

\item{ConsolidateOrder}{Do you want to "consolidate" order data. EG: 1 2 5 7-> 1 2 3 4. Useful if you intend to be removing some rows but... use with caution. NOTE: BEN, do I need to give more options here? Get function to only consolidate its own gaps perhaps?}

\item{RemoveMissingData}{Not currently implemented. General idea was to just drop lines whenever you have things like blank CODE, or Order==NA. Or you could have it flag these as concerns and pass them back perhaps.}
}
\value{
A new free list dataframe, without duplicates and such.
}
\description{
Given a Free list dataset with Codes, Ordering and subject numbers, tidy it up, detect duplicates, and just generally make it nicer.
}
\examples{
fakeData<- GenerateFakeFreeListData()
niceFakeData<- CalculateSalience(fakeData)
}
\keyword{Freelist}

