% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ComreyIterate.R
\name{ComreyIterate}
\alias{ComreyIterate}
\title{A Cat Function}
\usage{
ComreyIterate(M, A, MinZero = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

