% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ComreySolve.R
\name{ComreySolve}
\alias{ComreyIterate}
\alias{ComreySolve}
\title{ComreySolve}
\usage{
ComreySolve(M, precision = 0.005)
}
\arguments{
\item{M}{A square matrix.}

\item{precision}{The level of error at which the process will stop iterating and return an answer. Error is measured as "change in output vector between two iteration steps."}
}
\value{
A single vector of numbers representing the vector that was originally used to create M.
}
\description{
Given a square matrix, such that the matrix is the outer product of a single vector with itself (Except for the diagonal, which is ignored), determine the vector that was originally used.
ComreySolve also has a helper function called ComreyIterate. There isn't really any reason why you would use Comrey iterate by itself.
}
\note{
NOTE TO SELF: I really need to go back and re-read Comrey's paper to make sure I have done this right!
}
\examples{
x<- 1:5
mat<- x \\\%*\\\% t(x)
y<- ComreySolve(mat)
}
\author{
Alastair Jamieson-Lane
}
\references{
The Minimum Residual method of Factor Analysis.
Comrey, A.L.
Pychological reports, (1962) 11:15-18
}
\keyword{Consensus}

