% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ConsensusStressTest.R
\name{ConsensusStressTest}
\alias{ConsensusStressTest}
\title{ConsensusStressTest}
\usage{
ConsensusStressTest(numPeople, NumQuestions, numAns, Iterations)
}
\arguments{
\item{numPeople}{How many "People" do you wish to simulate answering each survey?}

\item{NumQuestions}{How many questions are there on each of your simulated surveys?}

\item{numAns}{How many possible answers are their to the questions on your virtual survey?}

\item{Iterations}{How many surveys do you wish to simulate?}
}
\value{
This function vector with two values: the number of errors, and the expected number of errors.
}
\description{
A function for simulating a large number of surveys, running consensus analysis on each, and then comparing the
true and calculated results to determine the algorithms error rate. Used to test \code{\link{ConsensusPipeline}}.
}
\note{
This function (and library) could probably use some additional features. If there are particular features you would like to see added, please email Jamieson-Lane, and he will see about adding them.
}
\examples{
ConsensusStressTest(15,15,4,500)
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@alumni.ubc.ca>
}
\keyword{Consensus}

