% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cultFST.R
\name{cultFST}
\alias{cultFST}
\title{cultFST}

\usage{
  cultFST(ni, nj, xi, xj)
}

\arguments{
  \item{ni}{Number of individuals in group i}
  \item{nj}{Number of individuals in group j}
  \item{xi}{Frequency of target trait in group i}
  \item{xj}{Frequency of target train in group j}
}

\value{
This function returns the Cultural FST value for two groups. If all of the variance lies between the groups, FST = 1. If all of the variance exists within groups, FST = 0.
  }

\description{
Cultural FST is a proportion of between-group cultural variance to the total cultural variance in a particular trait.}

\note{
This function is only to be applied to two groups. Use the fstmatrix() function for multiple groups.  
}

\examples{
  # Miscellaneous examples
  cultFST(100, 100, 100, 0) # n = 100 for both groups, i exhibits 100, j exhibits 0
  cultFST(100, 100, 50, 50) # n = 100 for both groups, i exhibits 50, j exhibits 50
  cultFST(100, 100, 0, 100) # n = 100 for both groups, i exhibits 0, j exhibits 100
  cultFST(100, 100, 46, 2) # n = 100 for both groups, i exhibits 46, j exhibits 2
  
  # Plot values of CF_ST across levels of xj and xi.
  par(mar = c(4, 4, 1, 1), mfrow = c(1, 1))
  plot(NA, xlim = c(0, 100), ylim = c(0, 1),
       xlab = expression(italic('x'['j'])), ylab = expression(italic('CF'['ST'])))
  lines(cultFST(100, 100, 0, 1:100), lty = 1, lwd = 1.5)
  lines(cultFST(100, 100, 25, 1:100), lty = 2, lwd = 1.5)
  lines(cultFST(100, 100, 50, 1:100), lty = 3, lwd = 1.5)
  lines(cultFST(100, 100, 75, 1:100), lty = 4, lwd = 1.5)
  lines(cultFST(100, 100, 100, 1:100), lty = 5, lwd = 1.5)
  legend(37, 1, c("0", "25", "50", "75", "100"), title = expression(italic('x'['i'])*' ='), 
         lty = 1:5, cex = .9)
}

\references{
  Bell, A. V., Richerson, P. J., & McElreath, R. (2009). Culture rather than genes provides greater scope for the evolution of large-scale human prosociality. Proceedings of the National Academy of Sciences, 106(42), 17671-17674.
  Handley, C., & Mathew, S. (2020). Human large-scale cooperation as a product of competition between cultural groups. Nature communications, 11(1), 702.
  Wright, S. (1965). The interpretation of population structure by F-statistics with special regard to systems of mating. Evolution, 395-420.
  }

\keyword{CulturalFST}