% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowerPlot.R
\name{FlowerPlot}
\alias{FlowerPlot}
\title{FlowerPlot}
\usage{
FlowerPlot(x, y)
}
\arguments{
\item{x}{This is the output of a SalienceByCode analysis (with Smith's S).}

\item{y}{This is the domain name you want with "quotes" around it (e.g., "Fruits").}
}
\value{
Returned will be a Flower Plot where the top item is the most salient item and Smith's S descends in a clockwise fashion. Connection weights' width correspond to Smith's S.
}
\description{
Make a flower plot (aka "BlumePlot" or "BlomsterPlot" or "ChechekPlot") automatically. If you can't contort it in-screen, then you'll have to manually do it (see code at https://anthrotools.wordpress.com). Sometimes your items will be too long to fit in the circles. Either change the data or--probably easier--is to rename the cells in the input table.
}
\note{
This graph will not be customizable without using code. Sometimes, items' names will be too long. You'll have to adjust accordingly if so. If you need to manually make one of these plots, try the code available at https://anthrotools.wordpress.com. This also uses a circle function that we found here: https://aurelienmadouasse.wordpress.com/2012/04/28/r-code-how-to-draw-a-circle/.
}
\examples{
data(FruitList)
FL.S <- CalculateSalience(FruitList)
model <- SalienceByCode(FL.S, CODE = "CODE", Subj = "Subj", Salience = "Salience", dealWithDoubles = "MAX")
par(mar=c(0,0,0,0))
FlowerPlot(model, "Fruits")

}
\keyword{FlowerPlot}
