% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/FreelistTable.MaxSal.R
\name{FreelistTable.MaxSal}
\alias{FreelistTable.MaxSal}
\title{A Cat Function}
\usage{
FreelistTable.MaxSal(mydata, CODE = "CODE", Salience = "Salience",
  Subj = "Subj", subjNum, CODEnum)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

