% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/FruitList.R
\docType{data}
\name{FruitList}
\alias{FruitList}
\title{FruitList}
\format{a \code{Date Frame}.}
\usage{
data(FruitList)
}
\description{
An example dataframe to illustrate what freelist data should look like.
}
\author{
Alastair Jamieson-Lane
}

