% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/GenerateConsensusData.R
\name{GenerateConsensusData}
\alias{GenerateConsensusData}
\title{A Cat Function}
\usage{
GenerateConsensusData(numPeople, numQuestions, numAns)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

