% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVSfaculty.R
\docType{data}
\name{RVSfaculty}
\alias{RVSfaculty}
\title{RVSfaculty}
\format{a \code{Date Frame}.}
\usage{
  data(RVSfaculty)
}
\description{
  An example dataframe to illustrate what triad data should look like as an input}.
}
\author{
    Benjamin Grant Purzycki. <bgpurzycki@cas.au.dk>
}