% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/SalienceByCode.R
\name{SalienceByCode}
\alias{SalienceByCode}
\title{A Cat Function}
\usage{
SalienceByCode(mydata, CODE = "CODE", Salience = "Salience",
  Subj = "Subj", dealWithDoubles = "DEFAULT")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

