\name{To Do list}
\alias{To Do list}

\title{To Do List}

\note{
Redo consensus pipeline and other consensus data stuff so that it can deal with "factor", or perhaps even strings as its input (currently it wants numbers, and that might be a bad thing) Would probably do this easiest by mapping to numbers, then doing stuff, then mapping back.

Make Consensus analysis sensitive to detecting "Bad datasets" - EG: does this dataset look like there are multiple domains of information? Does this dataset appear to suggest that some people have NEGATIVE expertise (and if so, how negative)... may need to read up more to do this, but it seems useful. Apparently the comrey iterator is NOT an especially good way of doing this. Will need to investigate.


Make Freelist method more robust. (Deal with empty entires, etc)

Make consensus capable of dealing with questions where each question has a different number of answers. (Is this mathematically feasible? I think so, but not sure.)
}



