% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad.gen.R
\name{triad.gen}
\alias{triad.gen}
\title{triad.gen}
\usage{
triad.gen(items)
}
\arguments{
  \item{items}{This is a vector of the items that will be in your triad test.
}
\value{
  The output here is a randomly generated set of triad tests. Simply export to a spreadsheet to incorporate into a survey.
  }
\description{
  You get a triad test instrument that's ready to export and use.
  }
\examples{
  RVS <- c("Ben", "Uffe", "Jesper", 
         "Lene", "Jorn", "Marianne Q.-F.",
         "Thomas", "Martin", "Mark") # your items
  triadmat <- triad.gen(RVS)
  View(triadmat)
}

\keyword{triad.gen}
