% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowerPlot.R
\name{triad.test}
\alias{triad.test}
\title{triad.test}
\usage{
  triad.test(data)
}
\arguments{
  \item{data}{This is triad data.}
}
\value{
  The output here is the similarity matrix of the triad data. 
  }
\description{
  You get a similarity matrix of triad test data.
}
\note{
  Pay close attention to the nature of the input data. The function will only work when data are in this particular shape and format. See the example below.
}
\examples{
  data(RVSfaculty)
  d.diff <- triad.test(RVSfaculty) # similarity matrix
  d.sim <- as.dist(1 - d.diff) # dissimilarity matrix
  plot(hclust(d.sim), main = NULL)
}
\keyword{triad.test}