% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/UglyList.R
\docType{data}
\name{UglyList}
\alias{UglyList}
\title{FruitList}
\format{a \code{Date Frame}.}
\usage{
data(UglyList)
}
\description{
An example dataframe to illustrate what free-list data might look like if things have gone not so smoothly. Used to illustrate the use of  \code{\link{CleanFreeList}}.
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@alumni.ubc.ca>
}

