% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldList.R
\docType{data}
\name{WorldList}
\alias{WorldList}
\title{WorldList}
\format{a \code{Date Frame}.}
\usage{
data(WorldList)
}
\description{
A bigger example dataframe to illustrate what free-list data with grouping should look like. If you want something smaller use \code{\link{FruitList}}.
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@cas.au.dk>
}

