% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CalculateSalience.R
\name{CalculateSalience}
\alias{CalculateSalience}
\title{A Cat Function}
\usage{
CalculateSalience(mydata, Order = "Order", Subj = "Subj", Norm = FALSE,
  Salience = "Salience")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

