% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CleanFreelist.R
\name{CleanFreeList}
\alias{CleanFreeList}
\title{CleanFreeList}
\usage{
CleanData<-CleanFreeList(mydata, Order="Order",Subj="Subj",CODE="CODE",ejectBadSubj=T,deleteDoubleCode=T,ConsolidateOrder=T,RemoveMissingData=T)
}
\arguments{
\item{mydata}{The free-list data. This should be a data frame, where each row contains a single response from a single respondent. For each such response, you need to know the subject number (or some form of unique identifier), the response (or "CODE" of the response and the ranking/order of the response (What was this respondents the first response? The second? The Seventeenth?). The "CODE" of the response is the target variable. We use "code" here as some free-list data gets coded into another coding scheme.}

\item{Order}{This is the name of the column which contains the "Order" information. For each subject responses should be ordered uniquely from 1 to N, where N is the number of responses. There should be no gaps or double ups. Defaults to "Order"}

\item{Subj}{This is the name of the column containing your subject names/numbers. Each subject should have a unique identifier in this column (Is this confusing, given that each subject may take several rows?). Defaults to "Subj".}

\item{CODE}{This is the name of the column containing your subject names/numbers. Each subject should have a unique identifier in this column (Is this confusing, given that each subject may take several rows?). Defaults to "Subj".}

\item{ejectBadSubj}{Do you want to eject all subjects who have (for whatever reason) bad data? For example duplicates or missing order entries. defaults to true.}

\item{deleteDoubleCode}{If someone says "Apple" twice, do you want to drop all but the first instance? defaults to true.}

\item{ConsolidateOrder}{Do you want to "consolidate" order data. EG: 1 2 5 7-> 1 2 3 4. Useful if you intend to be removing some rows.}

\item{RemoveMissingData}{Remove any row where Code is NA, or blank.}
}
\value{
A new free list dataframe, with all the requested error types removed. (NOTE, the effect that such removal will have on your statistics is unknown. We believe this method makes sensible modifications, but advise the use of caution.)
}
\description{
Given a Free list dataset with Codes, Ordering and subject numbers, remove incorrectly entered subjects, detect duplicate codes, remove blank codes, and enforce sequential ranking.
}
\examples{
data(UglyList)
View(CleanFreeList(UglyList))
View(CleanFreeList(UglyList),deleteDoubleCode=F)
}
\keyword{Freelist}

