% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ComreySolve.R
\name{ComreySolve}
\alias{ComreySolve}
\title{A Cat Function}
\usage{
ComreySolve(M, precision = 0.005)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

