% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ConsensusPipeline.R
\name{ConsensusPipeline}
\alias{ConsensusPipeline}
\title{ConsensusPipeline}
\usage{
ConsensusPipeline(SurveyResults, numQ)
}
\arguments{
\item{SurveyResults}{These are your survey results, written as a data frame. Rows are expected to represent participants, columns to represent individual questions. Each cell contains a particular individuals answer to a particular question- currently coded as a number (it is assumed things are somewhat multichoice).}

\item{numQ}{Currently this is a single number representing the number of possible answers to each question (hence for a true/false question, 2, for a multi choice, four perhaps).}
}
\value{
This function returns a list with four components.
\item{Answers}{a list of the functions estimated answers for each question.}
\item{Competance}{The estimated competance for each individual (the probability that they would KNOW the answer to some future question).}
\item{TestScore}{The supposed test score of each inidividual, assuming the answer key determined by the method}
\item{Probs}{the probability that each answer is correct for a given question ASSUMING that the method has correctly determined each individuals competence.
          Given how Competence is calculated, (from the model) it may be reckless to take this "probability" too seriously.}
}
\description{
The main function to use when trying to do consensus analysis on a selection of survey data. This function pretty much streamlines all the consensus analysis functions currently contained in the package into one function.
}
\details{
This method takes a set of survey results, along with a number representing the number of possible answers to each questions. It calculates the correlation between different peoples answers, corrects for random chance, estimate each individuals expertise and then determines the most likely answer for each question
}
\note{
This function (and library) could probably use some additional features. If there are particular feature's you would like to see added, please email me, and I will see about adding.
}
\examples{
FakeData<- GenerateConsensusData(8,8,4)
Survey <- FakeData$Survey
ConsensusResult <- ConsensusPipeline(Survey,4)
View(ConsensusResult$Answers)
View(FakeData$Answers)
}
\author{
Alastair Jamieson-Lane
}
\references{
Culture as consensus: a Theory of Culture and Informant Accuracy
A. Kimball Romney, Susan C. Weller, William H. Batchelder.
American Anthropologist, New Series, Volume 88, No 2 (Jun 1986) pp 313-338
}
\keyword{Consensus}

