% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ConsensusTestData.R
\docType{data}
\name{ConsensusTestData}
\alias{ConsensusTestData}
\title{ConsensusTestData}
\format{A \code{matrix}.}
\usage{
data(ConsensusTestData)
}
\description{
An example matrix of survey results should look when using \code{\link{ConsensusPipeline}}.
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@alumni.ubc.ca>
}

