% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreelistTable.freq.R
\name{FreeListTable.freq}
\alias{FreeListTable.MaxSal}
\alias{FreeListTable.Present}
\alias{FreeListTable.SumSal}
\alias{FreeListTable.freq}
\title{FreeListTable._______}
\usage{
FreeListTable.freq(mydata, CODE = "CODE", Salience = "Salience", Subj = "Subj", subjNum, CODEnum)
}
\arguments{
\item{mydata}{Your data, as a data frame, with rows for each response gathered. For each response, you are expected to have the CODE, salience and Subj.}

\item{CODE}{The Column of "mydata" that the CODE of the response can be found in.}

\item{Salience}{The column of "mydata" that the calculated Salience is to be found in. Note that this will have been determined earlier using "CalculateSalience"}

\item{Subj}{The column of "mydata" in which the subject number can be found.}

\item{subjNum}{The subject number of interest.}

\item{CODEnum}{The code you desire information about.}
}
\value{
A number, representing either: The presence/absence of the given code in the individuals responses (if "presence" was used). The total number of times the code was given in response (for "freq") The total Salience of a given code (summed over all responses with that code) The Maximum Salience (effectively the Salience levels when the Code was first used.)
}
\description{
This is a collection of helper functions for the function \code{\link{FreeListTable}}, that determine the relationship between a particular subject, and a particular code. For example, FreeListTable.freq counts how often an individual has responded with the specified Code. They are not intended for individual use.
}
\note{
If you are using these functions directly yourself, something has probably gone wrong.
}
\examples{
data(FruitList)
test<- FruitList
FreeListTable.freq(test,subjNum=7,CODEnum="apple") 
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@alumni.ubc.ca>
}
\keyword{FreeList}

