% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreelistTable.R
\name{FreeListTable}
\alias{FreeListTable}
\title{FreeListTable
 @aliases FreelistTable
Given a Free list data frame (with columns for subject number, order of responses and response code), produce a table comparing each subjects responses to each code. This table can be a simple "presence/absence" table for determining if a subject mentioned a particular code, or something more complex. \code{\link{colSums}} can be applied in order to get summaries (as opposed to person by person breakdown).}
\usage{
FreeListTable(mydata, CODE = "CODE", Salience = "Salience", Subj = "Subj", tableType = "DEFAULT")
}
\arguments{
\item{mydata}{This is your free-list data, stored as a data frame.}

\item{CODE}{The name of the column in which your "CODE" (eg, the subjects' individual responses) is stored.}

\item{Salience}{The name of the column in which each responses Salience is stored. If you wish to ask questions of salience, it is important you calculate salience before using this function. (using \code{\link{CalculateSalience}}).}

\item{Subj}{The name of the column where your subjects names/subject numbers are stored.}

\item{tableType}{Currently there are four types of tables: ÄúPRESENCE, úSUM_SALIENCE,MAX_SALIENCE and FREQUENCY. PRESENCEÄù will give a "1" if a participant mentioned the specified code, or Äú0Äù otherwise. If you specify ÄúFREQUENCYù, then you will get a count of how often each code was mentioned by each person. If you use ÄúSUM_SALIENCEÄù then you will get the total salience each person has associated with each code. If you use ÄúMAX_SALIENCEÄù then you will get the maximum salience, i.e., the salience of the code the first time it was mentioned.}

\item{GROUPING}{The name of the column where your subjects group names are sorted. Helps distinguish between individuals from different groups with the same ID number.}
}
\value{
The value returned is a data frame, where each row represents a subject, and each column (bar the first one or two) represents one of your free-list Codes. Depending on "tableType" the entries of the dataframe will either represent different things.
}
\description{
FreeListTable
 @aliases FreelistTable
Given a Free list data frame (with columns for subject number, order of responses and response code), produce a table comparing each subjects responses to each code. This table can be a simple "presence/absence" table for determining if a subject mentioned a particular code, or something more complex. \code{\link{colSums}} can be applied in order to get summaries (as opposed to person by person breakdown).
}
\examples{
fakeData<- GenerateFakeFreeListData() 
table<- FreeListTable(fakeData, tableType="PRESENCE")
View(table)
colSums(table) ##This will give summarised results.
data(WorldList)
FreeListTable(WorldList, tableType="FREQUENCY")
FreeListTable(WorldList, tableType="FREQUENCY",GROUPING="GROUPING")
WorldList<-CalculateSalience(WorldList,GROUPING="GROUPING")
FreeListTable(WorldList, tableType="SUM_SALIENCE",GROUPING="GROUPING")

}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@alumni.ubc.ca>
}
\keyword{FreeList}

