% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FruitList.R
\docType{data}
\name{FruitList}
\alias{FruitList}
\title{FruitList}
\format{a \code{Date Frame}.}
\usage{
data(FruitList)
}
\description{
An example dataframe to illustrate what free-list data should look like. If you want something with grouping use \code{\link{WorldList}}.
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@alumni.ubc.ca>
}

