% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/GenerateFakeFreelistData.R
\name{GenerateFakeFreeListData}
\alias{GenerateFakeFreeListData}
\title{GenerateFakeFreeListData}
\usage{
GenerateFakeFreeListData()
}
\value{
The function returns a single data frame, with columns "Subj",
 "Order" and "CODE". These column names are compatible with the
 default settings for the other FreeList related functions in this
  package. Subj specifies the subject number, CODE lists the subject's responses to a free-listing question (in this case "List whatever fruit you can think of") and Order specifies what order each subject listed their items in. The data created is entirely fictitious, and no attempt is made to prevent the imaginary subjects from listing the same fruit multiple times.
}
\description{
This function generates fake free-list data- both to act as a template, and incase anyone needs to run some sort of statistical analysis.
}
\note{
Depending on what is needed oruseful, we may implement a version of this function that can take some sort of statistical input (average length of peoples lists or such) and creates fake data with a particular shape.
This might be most easily implemented by having it use an existing data set and then mashing it up in some way so as to create a fake data set- allowing the user to compare their data to the usual in some way.
Doing this would require some amount of research, and thus I would only really want to do so if there were enough interest in it.
}
\examples{
fakeData<-GenerateFakeFreeListData()
View(fakeData)
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@alumni.ubc.ca>
}
\keyword{FreeList}

