% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadFromAnthropac.R
\name{LoadFromAnthropac}
\alias{LoadFromAnthropac}
\title{LoadFromAnthropac}
\usage{
LoadFromAnthropac(filename)
}
\arguments{
\item{filename}{This is the name of the file containing your anthropac data.}
}
\value{
A three column table containing the same data, stored in a format compatible with the AnthroTools package.
}
\description{
Given the file name of a file prepared in the anthropac format, load it in into a AnthroTools style table.
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@cas.au.dk>
}
\keyword{FreeList}

