% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/MakeAgreementMatrix.R
\name{MakeAgreementMatrix}
\alias{MakeAgreementMatrix}
\title{MakeAgreementMatrix}
\usage{
MakeAgreementMatrix(SurveyResults)
}
\arguments{
\item{SurveyResults}{A matrix containing the answers to a bunch of survey questions. Each row represents a particular individual, each column represents a particular question.}
}
\value{
A matrix M, where each entry represents the percentage of the time that individual i agreed with individual j.
}
\description{
Given a bunch of answers to a survey, this function makes a matrix representing what fraction of the time each person agreed with each other person.
}
\note{
This is probably already implemented elsewhere in R, and may be identical to a correlation matrix or something. I feel slightly silly putting it here to be honest. Oh well.
}
\examples{
FakeData<- GenerateConsensusData(8,8,4)
Survey <- FakeData$Survey
Ma <- MakeAgreementMatrix(Survey)
}
\author{
Alastair Jamieson-Lane
}
\keyword{Consensus}

