% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/MakeDiscountedAgreementMatrix.R
\name{MakeDiscountedAgreementMatrix}
\alias{MakeDiscountedAgreementMatrix}
\title{A Cat Function}
\usage{
MakeDiscountedAgreementMatrix(SurveyResults, numAns)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

