% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/MakeDiscountedAgreementMatrix.R
\name{MakeDiscountedAgreementMatrix}
\alias{MakeDiscountedAgreementMatrix}
\title{MakeDiscountedAgreementMatrix}
\usage{
MakeDiscountedAgreementMatrix(SurveyResults, numAns)
}
\arguments{
\item{SurveyResults}{A matrix containing the answers to a bunch of survey questions. Each row represents a particular individual and each column represents a particular question.}

\item{numAns}{This is a number, and represents the total number of answers available for each question. Currently it is assumed that all questions have the same number of answers. If I get enough requests, I'll look into the programming and mathematics required to vectorise this so that different questions can have different numbers of answers.}
}
\value{
A matrix, where each entry represents the probability that two individuals will both know the answer to some randomly selected question.
}
\description{
Given a bunch of answers to a survey, this function makes a matrix representing what fraction of the time each person agreed with each other person, and then discounts this value, based on how often you would expect this to happen by chance.
}
\examples{
FakeData<- GenerateConsensusData(8,8,4)
Survey <- FakeData$Survey
M <- MakeDiscountedAgreementMatrix(Survey, 4)
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Benjamin Grant Purzycki. <bgpurzycki@alumni.ubc.ca>#'
}
\keyword{Consensus}

