% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad.test.R
\name{triad.test}
\alias{triad.test}
\title{triad.test}
\usage{
triad.test(data)
}
\arguments{
  \item{data}{This is triad data. It needs to consist of: (1) an identifying number for each triad (NUM), (2-4) the specific items of the triad (item1, item2, item3), and (5-n) participants' selection of the most distinct item of the triad, the triad.test() function will create an item-by-item similarity matrix (PART1, PART2, ... PARTn).}
}
\value{
  The output here is the similarity matrix of the triad data. 
  }
\description{
  You get a similarity matrix of triad test data. You'll probably need to convert it to a dissimilarity matrix to plot it. See examples.
}
\note{
  Pay close attention to the nature of the input data. The function will only work when data are in this particular shape and format. See the example below.
}
\examples{
  data(RVSfaculty)
  d.diff <- triad.test(RVSfaculty) # similarity matrix
  d.sim <- as.dist(1 - d.diff) # dissimilarity matrix
  plot(hclust(d.sim), main = NULL)
}
\keyword{triad.test}
